Add noop image puller
An image puller is supposed to download a file, and to let podman know
where it downloaded it.
With macadam, we want to let the user specify which disk image they want
to use, and to run this.

This commit introduces a noop image puller which will be doing this, its
`Download` method does not do anything, it only tells podman to use the
file the user specified.