init: Add flags to specify ssh key and username
At the moment, due to the way podman starts gvproxy, it's mandatory to
be able to connect to the VM over SSH. Without working SSH, gvproxy will
exit after a few minutes leaving the VM without networking.

This commit adds a way to specify the SSH private key and usernname to
use as `macadam init` parameters.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>