convert qcow2 to raw on applehv
macadam uses vfkit in the backend and it only supports .iso and .raw images. However the RHEL KVM image published by RedHat are only in qcow2 format.
For this reason if someone wanted to run RHEL with macadam they needed to convert the image manually using qemu-img or other tools.

This patch adds the logic to perform that conversion automatically if a qcow2 image is fed to macadam.
Instead of copying the file as it is, it is converted to a .raw format and pushed to vfkit.

It uses go-qcow2reader from lima-vm and it leverages the convert dependency for a faster conversion -> https://github.com/lima-vm/lima/commit/0221c115332de50bb9f7d252a61498ddc5d72368

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>