remove validVM check as should be always true
When executing the SSH action macadam decides which username to use
based on the VM the command is targeting.
However when not defining a target machine, and so working with the default machine,
macadam uses the default "core" username. This can result in an error if the
username has been updated - like in WSL.

This is due to the variable validVM which is not set when macadam is not
specified a target machine and ends up using the existing default
VM.
This patch removes the !validVM check as it should never be true.
If the target machine does not exist it should fail earlier and if it
exists SSH command should use any username arg passed by the user or mc.SSH.RemoteUsername

Signed-off-by: lstocchi <lstocchi@redhat.com>