Update centos distro to qcow2
this commit updates the URL for fetching the CentOS Stream 10 image. The new URL points to a GenericCloud qcow2 image instead of the ec2 raw.xz image.
This is done for allowing running e2e test on ubuntu as the old image caused a disk space error.

It also removes the code to uncompress xz archives.

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>