allow to use local cloud-init config files to customize VM
it extends the init action to accept a list of cloud-init config files to customize the new VM.
The files must use the naming specified by cloud-init - user-data, meta-data and network-config.
One is free to pass any file. However, if user-data and meta-data are not passed or do not exist, a basic user-data will be generated to customize the VM.

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>